+++ R/read_labels_file.R
@@ -2,7 +2,7 @@
 #' @inheritParams default_params_doc
 #' @return a `labels_table`, as can be checked by \link{check_labels_table}
 #' @note `GCAE` does not save this data with column names.
-#' @seealso use \link{save_labels_file} to save a `labels_table`
+#' @seealso use \link{save_labels_table} to save a `labels_table`
 #' @examples
 #' labels_filename <- get_gcaer_filename("gcae_input_files_1_labels.csv")
 #' read_labels_file(labels_filename)