+++ scripts/upload_singularity_container.sh
@@ -17,11 +17,8 @@ echo "To sign the container, paste your passphrase"
 singularity sign gcaer.sif
 
 # Upload the latest version
-singularity push gcaer.sif library://richelbilderbeek/default/gcaer:0.6.2
+singularity push gcaer.sif library://richelbilderbeek/default/gcaer:0.6.3
 
 if [[ $HOSTNAME == "N141CU" ]]; then
   notify-send "Done uploading 'gcaer.sif'" "Done uploading 'gcaer.sif'"
 fi
-
-
-