+++ tests/testthat/test-save_labels_table.R
@@ -3,7 +3,7 @@ test_that("save and read must result in same table", {
     labels_filename = get_gcaer_filename("gcae_input_files_1_labels.csv")
   )
   labels_filename <- gcaer::get_gcaer_tempfilename(fileext = ".csv")
-  save_labels_file(
+  save_labels_table(
     labels_table = labels_table_before,
     labels_filename = labels_filename
   )